fix: Update stripe options input
