+++ packages/modules/payment/src/index.ts
@@ -2,14 +2,10 @@ import { ModuleExports } from "@medusajs/types"
 
 import { PaymentModuleService } from "@services"
 import loadProviders from "./loaders/providers"
-import loadDefaults from "./loaders/defaults"
 
-const service = PaymentModuleService
-const loaders = [loadProviders, loadDefaults] as any
-
-export const moduleDefinition: ModuleExports = {
-  service,
-  loaders,
+const moduleDefinition: ModuleExports = {
+  service: PaymentModuleService,
+  loaders: [loadProviders],
 }
 
 export default moduleDefinition
