+++ packages/modules/payment/src/loaders/index.ts
@@ -1,2 +1,1 @@
 export * from "./providers"
-export * from "./defaults"
