+++ packages/modules/payment/src/loaders/providers.ts
@@ -1,8 +1,14 @@
 import { moduleProviderLoader } from "@medusajs/modules-sdk"
-import { LoaderOptions, ModuleProvider, ModulesSdkTypes } from "@medusajs/types"
+import {
+  CreatePaymentProviderDTO,
+  LoaderOptions,
+  ModuleProvider,
+  ModulesSdkTypes,
+} from "@medusajs/types"
 import { Lifetime, asFunction, asValue } from "awilix"
 
 import * as providers from "../providers"
+import { PaymentProviderService } from "@services"
 
 const registrationFn = async (klass, container, pluginOptions) => {
   Object.entries(pluginOptions.config || []).map(([name, config]) => {
@@ -37,4 +43,32 @@ export default async ({
     providers: options?.providers || [],
     registerServiceFn: registrationFn,
   })
+
+  await registerProvidersInDb({ container })
+}
+
+const registerProvidersInDb = async ({
+  container,
+}: LoaderOptions): Promise<void> => {
+  const providersToLoad = container.resolve<string[]>("payment_providers")
+  const paymentProviderService = container.resolve<PaymentProviderService>(
+    "paymentProviderService"
+  )
+
+  const providers = await paymentProviderService.list({
+    id: providersToLoad,
+  })
+
+  const loadedProvidersMap = new Map(providers.map((p) => [p.id, p]))
+
+  const providersToCreate: CreatePaymentProviderDTO[] = []
+  for (const id of providersToLoad) {
+    if (loadedProvidersMap.has(id)) {
+      continue
+    }
+
+    providersToCreate.push({ id })
+  }
+
+  await paymentProviderService.create(providersToCreate)
 }
