+++ packages/modules/providers/payment-stripe/src/core/stripe-base.ts
@@ -23,7 +23,6 @@ import {
   ErrorCodes,
   ErrorIntentStatus,
   PaymentIntentOptions,
-  StripeCredentials,
   StripeOptions,
 } from "../types"
 import {
@@ -31,7 +30,7 @@ import {
   getSmallestUnit,
 } from "../utils/get-smallest-unit"
 
-abstract class StripeBase extends AbstractPaymentProvider<StripeCredentials> {
+abstract class StripeBase extends AbstractPaymentProvider<StripeOptions> {
   protected readonly options_: StripeOptions
   protected stripe_: Stripe
   protected container_: MedusaContainer
@@ -54,7 +53,7 @@ abstract class StripeBase extends AbstractPaymentProvider<StripeCredentials> {
 
   abstract get paymentIntentOptions(): PaymentIntentOptions
 
-  private validateOptions(options: StripeCredentials): void {
+  private validateOptions(options: StripeOptions): void {
     if (!isDefined(options.apiKey)) {
       throw new Error("Required option `apiKey` is missing in Stripe plugin")
     }
@@ -114,7 +113,7 @@ abstract class StripeBase extends AbstractPaymentProvider<StripeCredentials> {
     const { currency_code, amount } = input
 
     const description = (extra?.payment_description ??
-      this.options_?.payment_description) as string
+      this.options_?.paymentDescription) as string
 
     const intentRequest: Stripe.PaymentIntentCreateParams = {
       description,
@@ -125,7 +124,7 @@ abstract class StripeBase extends AbstractPaymentProvider<StripeCredentials> {
       ...intentRequestData,
     }
 
-    if (this.options_?.automatic_payment_methods) {
+    if (this.options_?.automaticPaymentMethods) {
       intentRequest.automatic_payment_methods = { enabled: true }
     }
